FileUtils: add function that clears the isSwap flag
There is a particular situation which requires that the FileUtils::isSwap
flag is unset when performing the read operation on an opened file. That
particular situation is when reading the magic number. Given the fact
that support for fat binaries is on its way, add a method to clear the
isSwap flag before a new read.