StringTable: add the offset of the current MachO as param
With the upcoming support of fat binaries, the StringTable now needs the
offset between the beginning of the file and the current architecture
(referred to as the current MachO inside of which the current
StringTable resides). The offset is needed during initialization, so
that it reads the file from the appropiate position.