MachO: add the offset member
Prior to implementing the Fat Binaries support, the MachO object needs
to be able to internally keep an offset inside the file itself. A Fat
Binary with two architectures will have two associated MachO objects,
each having its offset relative to the beginning of the file. This
offset marks the position where the MachHeader for that architecture
starts.