SymbolTableHeader: add file offset param
The Symbol Table Header needs to be aware of its position inside the
file, espacially in the case of Fat Binaries. The offset is needed for
subsequent fseek() calls, for example, where the exact position inside
the file needs to be known. Also update the getTableOffset() and
getStringTableOfset() calls with the newer getTableRealOffset() and
getStringTableRealOffset() that take into account the mentioned offset.